Url retreival and scrubbing functions.
The battlefield_stats.py script has been updated to retrieve the
webpage whose url is partly hard coded. Some fundamental error
checking has been implimented for the task.

A new directory has been added with the intention of it serving as a
module and it currently contains the initialization for the webpage
scrubbing function.