Formatting changes and scrub method fix.
The formatting changes made were to better comply with pep 8 coding
conventions and should therefore improve readability.

There was also an issue in the previous version of the code weap
method in the scrub module. The code contained some hardcoded references
to a tag in the webpage html that was infact not static and resulted in
the method breaking when used, nothing was being returned. Because the
use of the tag in the html was simply to filter out a dozen or so
results that weren't weapon stats the same result could be acheived by
simply ignoring them.