Bug fix for format check and play time conversion.
The bug, returning a false positive on xxs xxs (repeat entries) present in the
format check method has been fixed and now functions as desired. The addition
of a day field has also been added to the accepted format such that it is
now xxd xxh xxm xxs.

A method to convert the play time given in the format above into hours
has been created and functions as desired.