Edit of scrub module and completion of formatting.
The scrub module has been edited to change output a list again as it had
been before. This was to simplify the plotting function as extracting
information with a for loop is easier to do with a certain dictionary
format. The methods that do this formatting are under the plot module.

The formatting method is able to take the list from scrub and convert
it.