Implementation of a comparison plotting method.
A method that plots multiple profile stats on the same figure has been
developed. The implementation of this required a new method to be
created that adds a 0 value to stats for a weapon that is featured in
one profile but not another, due to the website not including weapons
the player hasn't used.

A method to generate plot parameters including color and x axis position
offset has been added to allow the bars to be displayed properly and
consistantly.

Things need doing:
  - addition of legend
  - completion of the main script using the funcitons developed
  - perhaps customization of color