Program completion.
The program has been successfully run and produces the desired plots.

Some additional arguments had to be added to the main script to
facilitate the inclusion of more than 1 profile to plot.

Some formatting had to be applied to the inclusion of text on the
figures including an if statement to control the number of decimal
places a float would be printed to. Additionally a legend was added to
the figures generated by s_c_comp_plot.

known issues:
  - the fillout method adds 0 to the corresponding entry in the other
     profiles, this however means that if a profile has 0.00 this result
     in having a mix of 0.00 and 0