More bug fixes and formatting changes
Changes made to the main script fixed an issue where the attempt
retrieve stats for an xbox account would fail with a 404 as the url
generation was faulty. The platform parameter was changed to xbl to fix
this.

Further formatting changes have been made to all modules to better
comply with pep 8