Overview comparison stats plotting complete.
The methods responsible for the scrubbing and formatting of the data
from the overview page of the battfield tracker website is complete.
These methods have then been successfully used to produce plots from the
newly developed over_comp_plot method run from the test script written for
it.

A bug fix was implemented for stats that consists of floating point
numbers. It turns out if these values get into the thousands a comma is
added to them which had to be removed. This was dicovered when using the
over_form method and converting a bugged (impossibly high) score/min value
into a float.