Major project restructure, formatting.
This version has undergone restructuring of the functions for the s_c_dicts.
The dictionary is now references as weap rather then s_c as this is the
information that it contains. The functions for handling the s_c dicts
have been moved out of the plot module and put into their own module
called weapons. The methods for handling of all dictionaries have also
been renamed so that they have more concise names. All test methods and the
main program have been updated to reflect this and have been verified to
work.

Additionally a minor change was made to the comparison plot functions to
sort the profile names in alphabetical order such that any other plots
created using the same user profiles will always be put in the same
directory.