Implementation of overview comparison and bug fix
The overview comparison method has been implemented in the main program
with supporting command line arguments.

There was a critical bug present in all previous versions of the
overview comparison method. The issue was that the x axis labels were
not set and nor were the y values being graphed. They instead were
referenced directly from the dictionary. This had the effect of having
the highest value for each profile plotted first, then the second
highest ect, resulting in a perato distribution for each soldier class
which was incorrect. Another issue was also present in the x axis
labelling. The result of this was that the last set of keys (soldier
classes) obtained from the dictionary were plotted in the order of
highest to lowest value. As such the last profile to be plotted had their
soldier classes plotted in descending order of value. This meant that the
only correct values on the entire figure were those for the last profile
plotted.
This was corrected with the implementation of a list of keys derived
from the master profile which was used to extract the y values as
well as set the y axis labels.