Creation of directory management module.
The major addition of this commit is the creation of the module
'directories' that will be used to create the file paths required for the
plotting functions. This will also facilitate the removal of all created
directories when the plotting functions fail.

Small edits were made to the main program and function doc string was
added to the overview comparison plot function.