Implementation of soldier class restrictions.
An option has been added to give the user control over the soldier
classes that are included in the overview stats plots.

A function to create a copy of an overview dictionary was added to allow
for the implementation of the above feature.

Additionally the directories functions have been renamed and configured
to support the creation of directories outside the pwd of the call to
python. There was a bug that prevented the creation of directories
outside the pwd as a complete directory reference for example, /home/...
meant that when splitting the name at each '/' char resulted in the
first directory created being ''. This was fixed by including an if
statement to change a '' on the split directory name to be '/'.